requirements.txt
python-telegram-bot==20.6
flask==2.3.3
gunicorn==21.2.0
